this can be removed
that is the way I did the click on it before. I didnt noticed merged but now it is as it is also merged the new way I coded to click on the button to continue logging in. it wont create any issues as it is after the continue and it is always skipped